test: fix docstring/websocket gaps and mirror test layout (#19, #20, #21)
- #19: add missing docstring to cleanup_globals._reset (docs-coverage 100%)
- #20: guard websocket route introspection against FastAPI _IncludedRouter
- #21: restructure tests/ to mirror src/pycharting/ so check_test_layout.py
  passes — flatten organizational grouping classes into module-level test
  functions, keep/add Test<SourceClass> classes (TestChartServer,
  TestDataManager, TestNoCacheStaticFiles, TestDataResponse, TestErrorResponse,
  TestDataValidationError), merge test_data_slicing into test_ingestion, fold
  test_import into test_interface.

Gates: fmt PASS, docs-coverage 100%, 148 tests PASS (coverage 100%), test-layout OK.
