+++ tests/pycharting/__init__.py
@@ -1,0 +1,1 @@
+"""Test suite mirroring the pycharting package layout."""
