--- tests/test_import.py
@@ -1,50 +1,0 @@
-"""Basic import tests for PyCharting package."""
-
-
-def test_package_import():
-    """Test that the main package can be imported."""
-    import pycharting
-
-    assert isinstance(pycharting.__version__, str)
-    assert pycharting.__version__
-
-
-def test_core_import():
-    """Test that core module can be imported."""
-    from pycharting import core
-
-    assert core is not None
-
-
-def test_data_import():
-    """Test that data module can be imported."""
-    from pycharting import data
-
-    assert data is not None
-
-
-def test_api_import():
-    """Test that api module can be imported."""
-    from pycharting import api
-
-    assert api is not None
-
-
-def test_web_import():
-    """Test that web module can be imported."""
-    from pycharting import web
-
-    assert web is not None
-
-
-def test_dependencies_available():
-    """Test that core dependencies are available."""
-    import fastapi
-    import numpy as np
-    import pandas as pd
-    import uvicorn
-
-    assert pd is not None
-    assert np is not None
-    assert fastapi is not None
-    assert uvicorn is not None
